refactor: replace app bar nav controllers with callbacks
