+++ app/src/main/java/io/joelt/texttemplate/ui/screens/draft_edit/DraftEditTopNavBar.kt
@@ -1,6 +1,5 @@
 package io.joelt.texttemplate.ui.screens.draft_edit
 
-import androidx.activity.compose.BackHandler
 import androidx.compose.material.IconButton
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Close
@@ -12,30 +11,13 @@ import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.navigation.NavHostController
 import io.joelt.texttemplate.R
-import io.joelt.texttemplate.navigation.Route
-import io.joelt.texttemplate.ui.screens.drafts.drafts
-
-fun NavHostController.navigateBackToDrafts() {
-    navigate(Route.drafts) {
-        this@navigateBackToDrafts.graph.startDestinationRoute?.let { route ->
-            popUpTo(route)
-        }
-    }
-}
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun DraftEditTopNavBar(nav: NavHostController, onSave: () -> Unit, onDelete: () -> Unit, onCopyToClipboard: () -> Unit) {
-    BackHandler(enabled = true) {
-        nav.navigateBackToDrafts()
-    }
-
+fun DraftEditTopNavBar(onBack: () -> Unit, onSave: () -> Unit, onDelete: () -> Unit, onCopyToClipboard: () -> Unit) {
     val backIconBtn = @Composable {
-        IconButton(onClick = {
-            nav.navigateBackToDrafts()
-        }) {
+        IconButton(onClick = onBack) {
             Icon(
                 imageVector = Icons.Default.Close,
                 contentDescription = stringResource(id = R.string.navbar_back),
