+++ app/src/main/java/io/joelt/texttemplate/ui/screens/settings/SettingsTopNavBar.kt
@@ -9,17 +9,14 @@ import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
-import androidx.navigation.NavHostController
 import io.joelt.texttemplate.R
 import io.joelt.texttemplate.ui.theme.Typography
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun SettingsTopNavBar(nav: NavHostController) {
+fun SettingsTopNavBar(onBack: () -> Unit) {
     val backIconBtn = @Composable {
-        IconButton(onClick = {
-            nav.popBackStack()
-        }) {
+        IconButton(onBack) {
             Icon(
                 imageVector = Icons.Default.ArrowBack,
                 contentDescription = stringResource(id = R.string.navbar_back),
